* src/gsad_omp.c (get_agents_omp): Change result_len to gsize.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/gsa@6237 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/gsad_omp.c
@@ -1229,7 +1229,7 @@ char *
 get_agents_omp (credentials_t * credentials,
                          const char * name,
                          const char * format,
-                         unsigned int *result_len,
+                         gsize *result_len,
                          const char * sort_field,
                          const char * sort_order)
 {
